[PC-425] refactor: core:excpetion 패키지 이름 수정 (org.yapp -> org.yapp.core.exception)
