+++ admin/src/main/java/org/yapp/auth/application/AuthService.java
@@ -8,8 +8,8 @@ import org.yapp.auth.application.dto.LoginDto;
 import org.yapp.core.auth.AuthToken;
 import org.yapp.core.auth.AuthTokenGenerator;
 import org.yapp.core.domain.user.User;
-import org.yapp.exception.ApplicationException;
-import org.yapp.exception.error.code.AuthErrorCode;
+import org.yapp.core.exception.ApplicationException;
+import org.yapp.core.exception.error.code.AuthErrorCode;
 import org.yapp.user.application.UserService;
 
 @Slf4j
