+++ admin/src/main/java/org/yapp/profile/application/AdminProfileService.java
@@ -8,8 +8,8 @@ import org.yapp.core.domain.profile.ProfileRejectHistory;
 import org.yapp.core.domain.profile.ProfileStatus;
 import org.yapp.core.domain.user.RoleStatus;
 import org.yapp.core.domain.user.User;
-import org.yapp.exception.ApplicationException;
-import org.yapp.exception.error.code.ProfileErrorCode;
+import org.yapp.core.exception.ApplicationException;
+import org.yapp.core.exception.error.code.ProfileErrorCode;
 import org.yapp.profile.dao.ProfileRejectHistoryRepository;
 import org.yapp.user.application.UserService;
