+++ api/src/main/java/org/yapp/domain/profile/application/ProfileService.java
@@ -13,6 +13,8 @@ import org.yapp.core.domain.profile.ProfileValuePick;
 import org.yapp.core.domain.profile.ProfileValueTalk;
 import org.yapp.core.domain.user.RoleStatus;
 import org.yapp.core.domain.user.User;
+import org.yapp.core.exception.ApplicationException;
+import org.yapp.core.exception.error.code.ProfileErrorCode;
 import org.yapp.domain.profile.dao.ProfileRepository;
 import org.yapp.domain.profile.presentation.request.ProfileBasicUpdateRequest;
 import org.yapp.domain.profile.presentation.request.ProfileCreateRequest;
@@ -21,8 +23,6 @@ import org.yapp.domain.profile.presentation.request.ProfileValuePickUpdateReques
 import org.yapp.domain.profile.presentation.request.ProfileValueTalkUpdateRequest;
 import org.yapp.domain.profile.presentation.request.ProfileValueTalkUpdateRequest.ProfileValueTalkPair;
 import org.yapp.domain.user.application.UserService;
-import org.yapp.exception.ApplicationException;
-import org.yapp.exception.error.code.ProfileErrorCode;
 
 @Service
 @RequiredArgsConstructor
