+++ api/src/main/java/org/yapp/domain/user/application/UserService.java
@@ -8,10 +8,10 @@ import org.yapp.core.auth.jwt.JwtUtil;
 import org.yapp.core.domain.profile.Profile;
 import org.yapp.core.domain.user.RoleStatus;
 import org.yapp.core.domain.user.User;
+import org.yapp.core.exception.ApplicationException;
+import org.yapp.core.exception.error.code.UserErrorCode;
 import org.yapp.domain.auth.presentation.dto.response.OauthLoginResponse;
 import org.yapp.domain.user.dao.UserRepository;
-import org.yapp.exception.ApplicationException;
-import org.yapp.exception.error.code.UserErrorCode;
 
 @Service
 @RequiredArgsConstructor
