+++ core/exception/src/main/java/org/yapp/core/exception/ApplicationException.java
@@ -1,8 +1,8 @@
-package org.yapp.exception;
+package org.yapp.core.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.yapp.exception.error.code.ErrorCode;
+import org.yapp.core.exception.error.code.ErrorCode;
 
 @Getter
 @RequiredArgsConstructor
