+++ core/exception/src/main/java/org/yapp/core/exception/error/code/ErrorCode.java
@@ -1,4 +1,4 @@
-package org.yapp.exception.error.code;
+package org.yapp.core.exception.error.code;
 
 import org.springframework.http.HttpStatus;
