+++ core/exception/src/main/java/org/yapp/core/exception/error/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package org.yapp.exception.error.response;
+package org.yapp.core.exception.error.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import java.util.List;
